+++ NEWS.md
@@ -1,3 +1,8 @@
+# SSNbler 1.1.1
+
+## Bug fixes
+* Fixed bug in `lines_to_lsn`, which prevented tonodes from snapping when the distance between them was snap_tolerance/1e4 < distance < snap_tolerance. Now, nodes are snapped if the distance between them is < snap_tolerance.
+
 # SSNbler 1.1.0
 
 ## Major updates